added entities file and made some models changes for the endpoints
