+++ tutto_api/__init__.py
@@ -1,0 +1,2 @@
+from tutto_api.src.models.endpoints import EndpointFactory  # noqa: F401
+from tutto_api.src.models.entities import Relative, Occurrence  # noqa: F401
